Put Variable class in its own header file
