+++ Calculator/Calculator.cpp
@@ -37,8 +37,7 @@ Input comes from cin through the Token_stream called ts.
 */
 
 #include "Facilities.h"
-
-
+#include "Variable.h"
 
 //------------------------------------------------------------------------------
 // Defines 3 Tokens of different types: a character token "Token(char ch)" with no value,
@@ -195,18 +194,6 @@ void Token_stream::ignore(char c)
         }
     }
 }
-
-//------------------------------------------------------------------------------
-// Creates a variable consisting of a string (which must start with an alpha character)
-// and a numeric value (which may be the result of an expression)
-
-class Variable {
-public:
-    string name;
-    double value;
-    Variable(string n, double val)
-        : name(n), value(val) {}
-};
 
 //------------------------------------------------------------------------------
 // A table for holding user declared variables
