+++ Calculator/Calculator.vcxproj.filters
@@ -23,5 +23,8 @@
     <ClInclude Include="Facilities.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="Variable.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
